Node and Statement to TS #355
