+++ package.json
@@ -22,6 +22,14 @@
     {
       "name": "Daniel Friedman",
       "url": "https://github.com/dan-f/"
+    },
+    {
+      "name": "Cénotélie",
+      "url": "https://github.com/cenotelie/"
+    },
+    {
+      "name": "Joep Meindertsma",
+      "url": "https://github.com/joepio/"
     }
   ],
   "license": "MIT",
