+++ src/default-graph.js
@@ -1,5 +1,5 @@
 'use strict'
-import Node from './node'
+import Node from './node-internal'
 
 export default class DefaultGraph extends Node {
   constructor () {
