+++ src/formula.js
@@ -6,7 +6,7 @@ import Literal from './literal'
 import log from './log'
 import NamedNode from './named-node'
 import Namespace from './namespace'
-import Node from './node'
+import Node from './node-internal'
 import Serializer from './serialize'
 import Statement from './statement'
 import Variable from './variable'
